requirements.txt
# Core web framework
fastapi>=0.110
uvicorn[standard]>=0.29

# Config & data model
pydantic>=2.6
pydantic-settings>=2.2

# HTTP client for proxy adapter
httpx>=0.27

# File uploads (FastAPI requires this to parse multipart/form-data)
python-multipart>=0.0.9

# Image processing used by Google adapter
Pillow>=10.0

# Google Gemini SDK (only在 PROVIDER=google 时使用)
google-genai>=0.3.0

